add result, context, timestamp, and stored files, remove corresponding tests
