+++ test/v1_0_3/non_templating.js
@@ -52,82 +52,6 @@
         });
     });
 
-    describe('An LRS returns a ContextActivity in an array, even if only a single ContextActivity is returned (4.1.6.2.c, 4.1.6.2.d)', function () {
-        var types = ['parent', 'grouping', 'category', 'other'];
-        this.timeout(0);
-
-        types.forEach(function (type) {
-            it('should return array for statement context "' + type + '"  when single ContextActivity is passed', function (done) {
-                var templates = [
-                    {statement: '{{statements.context}}'},
-                    {context: '{{contexts.' + type + '}}'}
-                ];
-                var data = createFromTemplate(templates);
-                data = data.statement;
-                data.id = helper.generateUUID();
-                var query = '?statementId=' + data.id;
-                var stmtTime = Date.now();
-                request(helper.getEndpointAndAuth())
-                    .post(helper.getEndpointStatements())
-                    .headers(helper.addAllHeaders({}))
-                    .json(data)
-                    .expect(200)
-                    .end()
-                    .get(helper.getEndpointStatements() + query)
-                    .wait(genDelay(stmtTime, query, data.id))
-                    .headers(helper.addAllHeaders({}))
-                    .expect(200)
-                    .end(function (err, res) {
-                        if (err) {
-                            done(err);
-                        } else {
-                            var statement = parse(res.body, done);
-                            expect(statement).to.have.property('context').to.have.property('contextActivities');
-                            expect(statement.context.contextActivities).to.have.property(type);
-                            expect(statement.context.contextActivities[type]).to.be.an('array');
-                            done();
-                        }
-                    });
-            });
-        });
-
-        types.forEach(function (type) {
-            it('should return array for statement substatement context "' + type + '"  when single ContextActivity is passed', function (done) {
-                var templates = [
-                    {statement: '{{statements.object_substatement}}'},
-                    {object: '{{substatements.context}}'},
-                    {context: '{{contexts.' + type + '}}'}
-                ];
-                var data = createFromTemplate(templates);
-                data = data.statement;
-                data.id = helper.generateUUID();
-                var query = '?statementId=' + data.id;
-                var stmtTime = Date.now();
-                request(helper.getEndpointAndAuth())
-                    .post(helper.getEndpointStatements())
-                    .headers(helper.addAllHeaders({}))
-                    .json(data)
-                    .expect(200)
-                    .end()
-                    .get(helper.getEndpointStatements() + query)
-                    .wait(genDelay(stmtTime, query, data.id))
-                    .headers(helper.addAllHeaders({}))
-                    .expect(200)
-                    .end(function (err, res) {
-                        if (err) {
-                            done(err);
-                        } else {
-                            var statement = parse(res.body, done);
-                            expect(statement).to.have.property('object').to.have.property('context').to.have.property('contextActivities');
-                            expect(statement.object.context.contextActivities).to.have.property(type);
-                            expect(statement.object.context.contextActivities[type]).to.be.an('array');
-                            done();
-                        }
-                    });
-            });
-        });
-    });
-
     describe('An LRS rejects with error code 400 Bad Request, a Request which uses Attachments and does not have a "Content-Type" header with value "application/json" or "multipart/mixed" (Format, 4.1.11.a, 4.1.11.b)', function () {
         var data;
         var attachment;
@@ -3241,121 +3165,6 @@
                         done();
                     }
                 });
-        });
-    });
-
-    describe('An LRS MUST accept statements with the stored property (Data 2.4.8.s3.b2)', function () {
-        this.timeout(0);
-        var storedTime = new Date('July 15, 2011').toISOString();
-        var template = [
-            {statement: '{{statements.default}}'},
-            {stored: storedTime}
-        ];
-        var data = createFromTemplate(template).statement;
-        var postId, putId;
-
-        it('using POST', function (done) {
-            var stmtTime = Date.now();
-            request(helper.getEndpointAndAuth())
-            .post(helper.getEndpointStatements())
-            .headers(helper.addAllHeaders())
-            .json(data)
-            .expect(200)
-            .end((err, res) => {
-                if (err) {
-                    done(err);
-                } else {
-                    postId = res.body[0];
-                    var query = '?statementId=' + postId;
-
-                    request(helper.getEndpointAndAuth())
-                    .get(helper.getEndpointStatements() + query)
-                    .wait(genDelay(stmtTime, query, postId))
-                    .headers(helper.addAllHeaders())
-                    .expect(200)
-                    .end((err, res) => {
-                        if (err) {
-                            done(err);
-                        } else {
-                            var result = parse(res.body);
-                            expect(result).to.have.property('stored');
-                            var stmtStored = result.stored;
-                            expect(stmtStored).to.not.eql(storedTime);
-                            done();
-                        }
-                    });
-                }
-            });
-        });
-
-        it('using PUT', function (done) {
-            putId = helper.generateUUID();
-            param = '?statementId=' + putId;
-            var stmtTime = Date.now();
-
-            request(helper.getEndpointAndAuth())
-            .put(helper.getEndpointStatements() + param)
-            .headers(helper.addAllHeaders())
-            .json(data)
-            .expect(204)
-            .end((err, res) => {
-                if (err) {
-                    done(err);
-                } else {
-                    request(helper.getEndpointAndAuth())
-                    .get(helper.getEndpointStatements() + param)
-                    .wait(genDelay(stmtTime, param, putId))
-                    .headers(helper.addAllHeaders())
-                    .expect(200)
-                    .end((err, res) => {
-                        if (err) {
-                            done(err);
-                        } else {
-                            var result = parse(res.body);
-                            expect(result).to.have.property('stored');
-                            var stmtStored = result.stored;
-                            expect(stmtStored).to.not.eql(storedTime);
-                            done();
-                        }
-                    });
-                }
-            });
-        });
-    });
-
-    describe('A stored property must be a TimeStamp (Data 2.4.8.s2)', function () {
-        it('retrieve statements, test a stored property', (done) => {
-            request(helper.getEndpointAndAuth())
-            .get(helper.getEndpointStatements())
-            .headers(helper.addAllHeaders())
-            .expect(200)
-            .end((err, res) => {
-                if (err) {
-                    done(err);
-                } else {
-                    var result = parse(res.body);
-                    var stmts = result.statements;
-                    var milliChecker = (num) => {
-                        expect(stmts[num]).to.have.property('stored');
-                        //formatted iso 8601
-                        var chkStored =  moment(stmts[num].stored, moment.ISO_8601);
-                        expect(chkStored.isValid()).to.be.true;
-                        expect(isNaN(chkStored._pf.parsedDateParts[6])).to.be.false;
-                        //precision to milliseconds
-                        if ((chkStored._pf.parsedDateParts[6] % 10) > 0) {
-                            expect(chkStored._pf.parsedDateParts[6] % 10).to.be.above(0);
-                            done();
-                        } else {
-                            if (++num < stmts.length) {
-                                milliChecker(num);
-                            } else {
-                                expect(chkStored._pf.parsedDateParts[6] % 10).to.be.above(0);
-                                done();
-                            }
-                        }
-                    }; milliChecker(0);
-                }
-            });
         });
     });
